+++ pkg/output/file_csv.go
@@ -9,7 +9,7 @@ import (
 	"github.com/ffuf/ffuf/pkg/ffuf"
 )
 
-var staticheaders = []string{"url", "redirectlocation", "position", "status_code", "content_length", "content_words", "content_lines", "resultfile"}
+var staticheaders = []string{"url", "redirectlocation", "position", "status_code", "content_length", "content_words", "content_lines", "content_type", "resultfile"}
 
 func writeCSV(config *ffuf.Config, res []Result, encode bool) error {
 	
@@ -68,6 +68,7 @@ func toCSV(r Result) []string {
 	res = append(res, strconv.FormatInt(r.ContentLength, 10))
 	res = append(res, strconv.FormatInt(r.ContentWords, 10))
 	res = append(res, strconv.FormatInt(r.ContentLines, 10))
+	res = append(res, r.ContentType)
 	res = append(res, r.ResultFile)
 	return res
 }