+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog
 - master
   - New
+     - All output file formats now include the `Content-Type`.
+     - New CLI flag `-recursion-strategy` that allows adding new queued recursion jobs for non-redirect responses.
   - Changed
   
 - v1.2.1